main.py
import argparse
import yaml
import torch
import os

from model.encoder import VAEEncoder
from model.generator import Generator
from model.discriminator import InfoGAN_Discriminator
from training.trainer import Trainer
from utils.helpers import set_seed
# Placeholder for data loading and tokenizer
from data.dataset import load_data, get_tokenizer 

def main(config_path: str):
    """Main function to run the paraphrase generation training pipeline."""
    # --- 1. Load Configuration --- 
    try:
        with open(config_path, 'r') as f:
            config = yaml.safe_load(f)
        print("Configuration loaded successfully:")
        # print(yaml.dump(config, indent=2))
    except FileNotFoundError:
        print(f"Error: Configuration file not found at {config_path}")
        return
    except yaml.YAMLError as e:
        print(f"Error parsing configuration file: {e}")
        return

    # --- 2. Set Seed for Reproducibility --- 
    seed = config.get('random_seed', 42)
    set_seed(seed)
    print(f"Random seed set to {seed}")

    # --- 3. Device Setup --- 
    device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
    print(f"Using device: {device}")

    # --- 4. Initialize Tokenizer (Placeholder) --- 
    # Replace with actual tokenizer loading based on config (e.g., from Hugging Face)
    tokenizer_name = config['data'].get('tokenizer_name', 'bert-base-uncased')
    tokenizer = get_tokenizer(tokenizer_name)
    vocab_size = tokenizer.vocab_size
    print(f"Tokenizer '{tokenizer_name}' loaded with vocab size: {vocab_size}")

    # --- 5. Load Data (Placeholder) --- 
    # Replace with actual data loading and preprocessing
    data_config = config['data']
    # This function should return dataloaders (train, val, test)
    # based on paths and batch size in data_config
    train_loader, val_loader, test_loader = load_data(data_config, tokenizer)
    print("Data loaders created (using placeholders). Actual data loading needed.")

    # --- 6. Initialize Models --- 
    model_config = config['model']
    
    # Assuming embed_dim is derived from encoder/generator structure or specified
    # Example: use a common embed_dim
    embed_dim = model_config.get('embed_dim', 768) # Example dimension
    
    encoder = VAEEncoder(
        input_dim=model_config['encoder'].get('input_dim', embed_dim), # Needs to match input embedding dim
        hidden_dim=model_config['encoder'].get('hidden_dim', 512),
        latent_dim=model_config['encoder'].get('latent_dim', 256)
    ).to(device)
    print("VAEEncoder initialized.")

    generator = Generator(
        embed_dim=embed_dim, # Must match VAE latent dim if directly connected, or decoder input dim
        hidden_dim=model_config['generator'].get('hidden_dim', 2048),
        vocab_size=vocab_size, # From tokenizer
        num_layers=model_config['generator'].get('num_layers', 6),
        nhead=model_config['generator'].get('nhead', 8),
        dropout=model_config['generator'].get('dropout', 0.1),
        max_seq_len=data_config.get('max_seq_len', 100)
    ).to(device)
    print("Generator initialized.")

    discriminator = InfoGAN_Discriminator(
        # Input dim depends on how condition (z) and paraphrase are combined
        # Example: concat(z_enc, paraphrase_embedding_pooled)
        input_dim=model_config['encoder'].get('latent_dim', 256) + embed_dim, # Placeholder combination
        hidden_dim=model_config['discriminator'].get('hidden_dim', 1024),
        disc_code_dim=model_config['discriminator'].get('infogan_discrete_code_dim', 10),
        cont_code_dim=model_config['discriminator'].get('infogan_continuous_code_dim', 2),
        mbd_out_features=model_config['discriminator'].get('mbd_out_features', 128),
        mbd_kernel_dim=model_config['discriminator'].get('mbd_kernel_dim', 64)
    ).to(device)
    print("InfoGAN_Discriminator initialized.")

    # --- 7. Initialize Trainer --- 
    trainer = Trainer(encoder, generator, discriminator, config)
    print("Trainer initialized.")

    # --- 8. Start Training --- 
    print("Starting training...")
    trainer.train(train_loader, val_loader)

    # --- 9. Optional: Final Evaluation --- 
    # print("Training finished. Running final evaluation on test set...")
    # test_results = evaluate_model(generator, encoder, test_loader, config['evaluation'], device, tokenizer)
    # print("Test Results:", test_results)

if __name__ == '__main__':
    parser = argparse.ArgumentParser(description='Train a VAE-InfoGAN Paraphrase Model.')
    parser.add_argument('--config', type=str, default='configs/default.yaml', 
                        help='Path to the configuration YAML file.')
    args = parser.parse_args()
    
    main(args.config)
